requirements.txt
dify_plugin>=0.2.0,<0.3.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0